+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Improvements:
 * [OLMIS-5134](https://openlmis.atlassian.net/browse/OLMIS-5134): Added ability to find user's role assignments by supervisoryNodeId and programId.
 * [OLMIS-5409](https://openlmis.atlassian.net/browse/OLMIS-5409): Updated ui-components to version 7.0.0.
 * [OLMIS-3696](https://openlmis.atlassian.net/browse/OLMIS-3696): Added dependency and development dependency locking.
+* [OLMIS-5415](https://openlmis.atlassian.net/browse/OLMIS-5415): Disabled edit of some fields for externally managed facilities.
 
 5.4.1 / 2018-10-01
 ==================